# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "AI股票助手API",
    "description": "AI股票助手后端API接口文档",
    "version": "1.0.0",
    "contact": {
      "name": "AI股票助手开发团队"
    }
  },
  "host": "localhost:3033",
  "basePath": "/api",
  "schemes": ["http", "https"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
...
    "/quotes": {
      "post": {
        "tags": ["行情数据"],
        "summary": "创建行情快照",
        "description": "创建单个行情快照记录",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateQuoteDto"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "创建成功",
            "schema": {
              "$ref": "#/definitions/Quote"
            }
          }
        }
      },
      "get": {
        "tags": ["行情数据"],
        "summary": "获取所有行情快照",
        "description": "根据查询条件获取行情快照列表",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "type": "string",
            "description": "股票代码"
          },
          {
            "name": "market",
            "in": "query",
            "type": "string",
            "description": "市场"
          },
          {
            "name": "marketCode",
            "in": "query",
            "type": "string",
            "description": "市场代码"
          },
          {
            "name": "startTime",
            "in": "query",
            "type": "string",
            "format": "date-time",
            "description": "开始时间"
          },
          {
            "name": "endTime",
            "in": "query",
            "type": "string",
            "format": "date-time",
            "description": "结束时间"
          },
          {
            "name": "page",
            "in": "query",
            "type": "integer",
            "description": "页码"
          },
          {
            "name": "limit",
            "in": "query",
            "type": "integer",
            "description": "每页数量"
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Quote"
              }
            }
          }
        }
      }
    },
    "/quotes/batch": {
      "post": {
        "tags": ["行情数据"],
        "summary": "批量创建行情快照",
        "description": "批量创建多个行情快照记录",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/CreateQuoteDto"
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "批量创建成功",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Quote"
              }
            }
          }
        }
      }
    },
    "/quotes/{id}": {
      "get": {
        "tags": ["行情数据"],
        "summary": "根据ID获取行情快照",
        "description": "根据ID获取单个行情快照",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "行情快照ID"
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "schema": {
              "$ref": "#/definitions/Quote"
            }
          }
        }
      },
      "put": {
        "tags": ["行情数据"],
        "summary": "更新行情快照",
        "description": "更新指定的行情快照",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "行情快照ID"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateQuoteDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "更新成功",
            "schema": {
              "$ref": "#/definitions/Quote"
            }
          }
        }
      },
      "delete": {
        "tags": ["行情数据"],
        "summary": "删除行情快照",
        "description": "删除指定的行情快照",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "行情快照ID"
          }
        ],
        "responses": {
          "204": {
            "description": "删除成功"
          }
        }
      }
    },
    "/quotes/latest/{code}": {
      "get": {
        "tags": ["行情数据"],
        "summary": "获取指定股票的最新行情",
        "description": "获取指定股票代码的最新行情快照",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "股票代码"
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "schema": {
              "$ref": "#/definitions/Quote"
            }
          }
        }
      }
    },
    "/quotes/history/{code}": {
      "get": {
        "tags": ["行情数据"],
        "summary": "获取指定股票的历史行情",
        "description": "获取指定股票的历史行情快照",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "股票代码"
          },
          {
            "name": "startTime",
            "in": "query",
            "type": "string",
            "format": "date-time",
            "description": "开始时间"
          },
          {
            "name": "endTime",
            "in": "query",
            "type": "string",
            "format": "date-time",
            "description": "结束时间"
          },
          {
            "name": "limit",
            "in": "query",
            "type": "integer",
            "description": "限制数量"
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Quote"
              }
            }
          }
        }
      }
    },
    "/quotes/date/{code}": {
      "get": {
        "tags": ["行情数据"],
        "summary": "根据日期获取指定股票的行情",
        "description": "根据日期范围获取指定股票的行情快照",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "股票代码"
          },
          {
            "name": "startDate",
            "in": "query",
            "type": "string",
            "format": "date",
            "description": "开始日期"
          },
          {
            "name": "endDate",
            "in": "query",
            "type": "string",
            "format": "date",
            "description": "结束日期"
          },
          {
            "name": "limit",
            "in": "query",
            "type": "integer",
            "description": "限制数量"
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Quote"
              }
            }
          }
        }
      }
    },
    "/quotes/date": {
      "get": {
        "tags": ["行情数据"],
        "summary": "获取指定日期的所有行情",
        "description": "获取指定日期的所有股票行情快照",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": true,
            "type": "string",
            "format": "date",
            "description": "日期"
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Quote"
              }
            }
          }
        }
      }
    },
    "/quotes/range": {
      "delete": {
        "tags": ["行情数据"],
        "summary": "批量删除指定时间范围的行情快照",
        "description": "删除指定时间范围内的所有行情快照",
        "parameters": [
          {
            "name": "startTime",
            "in": "query",
            "required": true,
            "type": "string",
            "format": "date-time",
            "description": "开始时间"
          },
          {
            "name": "endTime",
            "in": "query",
            "required": true,
            "type": "string",
            "format": "date-time",
            "description": "结束时间"
          }
        ],
        "responses": {
          "204": {
            "description": "删除成功"
          }
        }
      }
    },
...
    "/quotes/rankings/gainers": {
      "get": {
        "tags": ["行情数据"],
        "summary": "获取涨跌幅排行榜",
        "description": "获取涨幅最大的股票排行榜",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "type": "integer",
            "description": "限制数量"
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Quote"
              }
            }
          }
        }
      }
    },
    "/quotes/rankings/losers": {
      "get": {
        "tags": ["行情数据"],
        "summary": "获取跌幅排行榜",
        "description": "获取跌幅最大的股票排行榜",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "type": "integer",
            "description": "限制数量"
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Quote"
              }
            }
          }
        }
      }
    },
    "/quotes/rankings/volume": {
      "get": {
        "tags": ["行情数据"],
        "summary": "获取成交量排行榜",
        "description": "获取成交量最大的股票排行榜",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "type": "integer",
            "description": "限制数量"
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Quote"
              }
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Quote": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "description": "行情快照ID"
        },
        "code": {
          "type": "string",
          "description": "股票代码"
        },
        "name": {
          "type": "string",
          "description": "股票名称"
        },
        "market": {
          "type": "string",
          "description": "市场"
        },
        "marketCode": {
          "type": "string",
          "description": "市场代码"
        },
        "pe": {
          "type": "number",
          "format": "decimal",
          "description": "市盈率"
        },
        "latestPrice": {
          "type": "number",
          "format": "decimal",
          "description": "最新价"
        },
        "changePercent": {
          "type": "number",
          "format": "decimal",
          "description": "涨跌幅(%)"
        },
        "changeAmount": {
          "type": "number",
          "format": "decimal",
          "description": "涨跌额"
        },
        "openPrice": {
          "type": "number",
          "format": "decimal",
          "description": "开盘价"
        },
        "highPrice": {
          "type": "number",
          "format": "decimal",
          "description": "最高价"
        },
        "lowPrice": {
          "type": "number",
          "format": "decimal",
          "description": "最低价"
        },
        "volume": {
          "type": "integer",
          "format": "int64",
          "description": "成交量(股)"
        },
        "volumeAmount": {
          "type": "number",
          "format": "decimal",
          "description": "成交额(元)"
        },
        "previousClosePrice": {
          "type": "number",
          "format": "decimal",
          "description": "昨收价"
        },
        "snapshotTime": {
          "type": "string",
          "format": "date-time",
          "description": "快照时间"
        },
        "snapshotDate": {
          "type": "string",
          "format": "date",
          "description": "快照日期"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time",
          "description": "创建时间"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time",
          "description": "更新时间"
        }
      }
    },
...
    "CreateQuoteDto": {
      "type": "object",
      "required": ["code", "name", "market", "marketCode"],
      "properties": {
        "code": {
          "type": "string",
          "description": "股票代码"
        },
        "name": {
          "type": "string",
          "description": "股票名称"
        },
        "market": {
          "type": "string",
          "description": "市场"
        },
        "marketCode": {
          "type": "string",
          "description": "市场代码"
        },
        "pe": {
          "type": "number",
          "format": "decimal",
          "description": "市盈率"
        },
        "latestPrice": {
          "type": "number",
          "format": "decimal",
          "description": "最新价"
        },
        "changePercent": {
          "type": "number",
          "format": "decimal",
          "description": "涨跌幅"
        },
        "changeAmount": {
          "type": "number",
          "format": "decimal",
          "description": "涨跌额"
        },
        "openPrice": {
          "type": "number",
          "format": "decimal",
          "description": "开盘价"
        },
        "highPrice": {
          "type": "number",
          "format": "decimal",
          "description": "最高价"
        },
        "lowPrice": {
          "type": "number",
          "format": "decimal",
          "description": "最低价"
        },
        "volume": {
          "type": "integer",
          "format": "int64",
          "description": "成交量"
        },
        "volumeAmount": {
          "type": "number",
          "format": "decimal",
          "description": "成交额"
        },
        "previousClosePrice": {
          "type": "number",
          "format": "decimal",
          "description": "昨收价"
        },
        "snapshotTime": {
          "type": "string",
          "format": "date-time",
          "description": "快照时间"
        },
        "snapshotDate": {
          "type": "string",
          "format": "date",
          "description": "快照日期"
        }
      }
    },
    "UpdateQuoteDto": {
      "type": "object",
      "properties": {
        "pe": {
          "type": "number",
          "format": "decimal",
          "description": "市盈率"
        },
        "latestPrice": {
          "type": "number",
          "format": "decimal",
          "description": "最新价"
        },
        "changePercent": {
          "type": "number",
          "format": "decimal",
          "description": "涨跌幅"
        },
        "changeAmount": {
          "type": "number",
          "format": "decimal",
          "description": "涨跌额"
        },
        "openPrice": {
          "type": "number",
          "format": "decimal",
          "description": "开盘价"
        },
        "highPrice": {
          "type": "number",
          "format": "decimal",
          "description": "最高价"
        },
        "lowPrice": {
          "type": "number",
          "format": "decimal",
          "description": "最低价"
        },
        "volume": {
          "type": "integer",
          "format": "int64",
          "description": "成交量"
        },
        "volumeAmount": {
          "type": "number",
          "format": "decimal",
          "description": "成交额"
        },
        "previousClosePrice": {
          "type": "number",
          "format": "decimal",
          "description": "昨收价"
        },
        "snapshotTime": {
          "type": "string",
          "format": "date-time",
          "description": "快照时间"
        },
        "snapshotDate": {
          "type": "string",
          "format": "date",
          "description": "快照日期"
        }
      }
    },
...
  },
...
}